+++ lua/codecompanion/strategies/chat/agents/tools/cmd_runner.lua
@@ -5,26 +5,21 @@ commands in the same XML block. All commands must be approved by you.
 --]]
 
 local config = require("codecompanion.config")
-
-local log = require("codecompanion.utils.log")
 local util = require("codecompanion.utils")
-local xml2lua = require("codecompanion.utils.xml.xml2lua")
-
----@class CmdRunner.ChatOpts
----@field cmd table|string The command that was executed
----@field output table|string The output of the command
----@field message? string An optional message
 
 ---Outputs a message to the chat buffer that initiated the tool
 ---@param msg string The message to output
----@param tool CodeCompanion.Tools The tools object
----@param opts CmdRunner.ChatOpts
-local function to_chat(msg, tool, opts)
-  if type(opts.cmd) == "table" then
-    opts.cmd = table.concat(opts.cmd, " ")
+---@param agent CodeCompanion.Agent The tools object
+---@param opts {cmd: table, output: table|string, message?: string}
+local function to_chat(msg, agent, opts)
+  local cmd
+  if opts and type(opts.cmd) == "table" then
+    cmd = table.concat(opts.cmd, " ")
+  else
+    cmd = opts.cmd
   end
-  if type(opts.output) == "table" then
-    opts.output = table.concat(opts.output, "\n")
+  if opts and type(opts.output) == "table" then
+    opts.output = vim.iter(opts.output):flatten():join("\n")
   end
 
   local content
@@ -34,7 +29,7 @@ local function to_chat(msg, tool, opts)
 
 ]],
       msg,
-      opts.cmd
+      cmd
     )
   else
     content = string.format(
@@ -46,18 +41,18 @@ local function to_chat(msg, tool, opts)
 
 ]],
       msg,
-      opts.cmd,
+      cmd,
       opts.output
     )
   end
 
-  return tool.chat:add_buf_message({
+  return agent.chat:add_buf_message({
     role = config.constants.USER_ROLE,
     content = content,
   })
 end
 
----@class CodeCompanion.Tool
+---@class CodeCompanion.Agent.Tool
 return {
   name = "cmd_runner",
   cmds = {
@@ -93,7 +88,7 @@ return {
       },
     },
   },
-  system_prompt = function(schema)
+  system_prompt = function(schema, xml2lua)
     return string.format(
       [[## Command Runner Tool (`cmd_runner`) – Enhanced Guidelines
 
@@ -161,9 +156,9 @@ return {
     )
   end,
   handlers = {
-    ---@param self CodeCompanion.Tools The tool object
-    setup = function(self)
-      local tool = self.tool --[[@type CodeCompanion.Tool]]
+    ---@param agent CodeCompanion.Agent The tool object
+    setup = function(agent)
+      local tool = agent.tool --[[@type CodeCompanion.Agent.Tool]]
       local action = tool.request.action
       local actions = vim.isarray(action) and action or { action }
 
@@ -175,49 +170,50 @@ return {
         table.insert(tool.cmds, entry)
       end
     end,
-
-    ---Approve the command to be run
-    ---@param self CodeCompanion.Tools The tool object
-    ---@param cmd table
-    ---@return boolean
-    approved = function(self, cmd)
-      if vim.g.codecompanion_auto_tool_mode then
-        log:info("[Cmd Runner Tool] Auto-approved running the command")
-        return true
-      end
-
-      local cmd_concat = table.concat(cmd.cmd or cmd, " ")
-
-      local msg = "Run command: `" .. cmd_concat .. "`?"
-      local ok, choice = pcall(vim.fn.confirm, msg, "No\nYes")
-      if not ok or choice ~= 2 then
-        log:info("[Cmd Runner Tool] Rejected running the command")
-        return false
-      end
-
-      log:info("[Cmd Runner Tool] Approved running the command")
-      return true
-    end,
   },
 
   output = {
+    ---The message which is shared with the user when asking for their approval
+    ---@param agent CodeCompanion.Agent
+    ---@param self CodeCompanion.Agent.Tool
+    ---@return string
+    prompt = function(agent, self)
+      local cmds = self.cmds
+      if vim.tbl_count(cmds) == 1 then
+        return string.format("Run the command `%s`?", table.concat(cmds[1].cmd, " "))
+      end
+
+      local individual_cmds = vim.tbl_map(function(c)
+        return table.concat(c.cmd, " ")
+      end, cmds)
+      return string.format("Run the following commands?\n\n%s", table.concat(individual_cmds, "\n"))
+    end,
+
     ---Rejection message back to the LLM
-    rejected = function(self, cmd)
-      to_chat("I chose not to run", self, { cmd = cmd.cmd or cmd, output = "" })
+    ---@param agent CodeCompanion.Agent
+    ---@param cmd table
+    ---@return nil
+    rejected = function(agent, cmd)
+      to_chat("I chose not to run", agent, { cmd = cmd.cmd or cmd, output = "" })
     end,
 
-    ---@param self CodeCompanion.Tools The tools object
-    ---@param cmd table|string The command that was executed
-    ---@param stderr table|string
-    error = function(self, cmd, stderr)
-      to_chat("There was an error from", self, { cmd = cmd.cmd or cmd, output = stderr })
+    ---@param agent CodeCompanion.Agent
+    ---@param cmd table
+    ---@param stderr table
+    ---@param stdout? table
+    error = function(agent, cmd, stderr, stdout)
+      to_chat("There was an error from", agent, { cmd = cmd.cmd or cmd, output = stderr })
+
+      if stdout and not vim.tbl_isempty(stdout) then
+        to_chat("There was also some output from", agent, { cmd = cmd.cmd or cmd, output = stdout })
+      end
     end,
 
-    ---@param self CodeCompanion.Tools The tools object
-    ---@param cmd table|string The command that was executed
-    ---@param stdout table|string
-    success = function(self, cmd, stdout)
-      to_chat("The output from", self, { cmd = cmd.cmd or cmd, output = stdout })
+    ---@param agent CodeCompanion.Agent
+    ---@param cmd table The command that was executed
+    ---@param stdout table
+    success = function(agent, cmd, stdout)
+      to_chat("The output from", agent, { cmd = cmd.cmd or cmd, output = stdout })
     end,
   },
 }